refactor(#3053): move CBs to fim.event_monitor
